refactor: Key Vault Child -> Azure package
